rewrites AsyncCoreServerSpec to CoreServerSpec
