+++ src/test/scala/viper/server/core/AstGenerationTests.scala
@@ -9,13 +9,14 @@ package viper.server.core
 import java.nio.file.NoSuchFileException
 
 import akka.http.scaladsl.testkit.ScalatestRouteTest
-import org.scalatest.{Matchers, WordSpec}
+import org.scalatest.matchers.should.Matchers
+import org.scalatest.wordspec.AnyWordSpec
 import viper.server.utility.AstGenerator
 import viper.silver.ast.Program
 import viper.silver.logger.ViperStdOutLogger
 
 
-class AstGenerationTests extends WordSpec with Matchers with ScalatestRouteTest {
+class AstGenerationTests extends AnyWordSpec with Matchers with ScalatestRouteTest {
 
   private val verifiableFile = "src/test/resources/viper/let.vpr"
   private val emptyFile = "src/test/resources/viper/empty.vpr"
